moved campaign class
